ptn test refactor: move common functions to support
