+++ test/test_ohClassPtn_appRibPtn.cpp
@@ -18,51 +18,6 @@ struct CSR {
   redev::GOs val;  
 };
 
-void prepareAppOutMessage(Omega_h::Mesh& mesh, const redev::ClassPtn& ptn,
-    ts::OutMsg& out, redev::LOs& permute) {
-  //transfer vtx classification to host
-  auto classIds = mesh.get_array<Omega_h::ClassId>(0, "class_id");
-  auto classIds_h = Omega_h::HostRead(classIds);
-  //count number of vertices going to each destination process by calling getRank - degree array
-  std::map<int,int> destRankCounts;
-  const auto ptnRanks = ptn.GetRanks();
-  for(auto rank : ptnRanks) {
-    destRankCounts[rank] = 0;
-  }
-  for(auto i=0; i<classIds_h.size(); i++) {
-    auto dr = ptn.GetRank(classIds_h[i]);
-    assert(destRankCounts.count(dr));
-    destRankCounts[dr]++;
-  }
-
-  //create dest and offsets arrays from degree array
-  out.offset.resize(destRankCounts.size()+1);
-  out.dest.resize(destRankCounts.size());
-  out.offset[0] = 0;
-  int i = 1;
-  for(auto rankCount : destRankCounts) {
-    out.dest[i-1] = rankCount.first;
-    out.offset[i] = out.offset[i-1]+rankCount.second;
-    i++;
-  }
-
-  //fill permutation array such that for vertex i permute[i] contains the
-  //  position of vertex i's data in the message array
-  std::map<int,int> destRankIdx;
-  for(size_t i=0; i<out.dest.size(); i++) {
-    auto dr = out.dest[i];
-    destRankIdx[dr] = out.offset[i];
-  }
-  auto gids = mesh.globals(0);
-  auto gids_h = Omega_h::HostRead(gids);
-  permute.resize(classIds_h.size());
-  for(auto i=0; i<classIds_h.size(); i++) {
-    auto dr = ptn.GetRank(classIds_h[i]);
-    auto idx = destRankIdx[dr]++;
-    permute[i] = idx;
-  }
-}
-
 //from https://stackoverflow.com/a/12399290
 template <typename T>
 std::vector<size_t> sort_indexes(const T &v) {
@@ -170,12 +125,6 @@ void getRdvPermutation(Omega_h::Mesh& mesh, const redev::GOs& inGids, redev::GOs
   }
 }
 
-void writeVtk(Omega_h::Mesh& mesh, std::string name, int step) {
-  std::stringstream ss;
-  ss << name << step << ".vtk";
-  Omega_h::vtk::write_parallel(ss.str(), &mesh, mesh.dim());
-}
-
 int main(int argc, char** argv) {
   auto lib = Omega_h::Library(&argc, &argv);
   auto world = lib.world();
@@ -199,11 +148,11 @@ int main(int argc, char** argv) {
     ts::getClassPtn(mesh, ranks, classIds);
     REDEV_ALWAYS_ASSERT(ranks.size()==3);
     REDEV_ALWAYS_ASSERT(ranks.size()==classIds.size());
-    Omega_h::vtk::write_parallel("rdvSplit.vtk", &mesh, mesh.dim());
+    ts::writeVtk(mesh,"rdvSplit",0);
   } else {
     REDEV_ALWAYS_ASSERT(world->size()==2);
     if(!rank) REDEV_ALWAYS_ASSERT(mesh.nelems()==11);
-    Omega_h::vtk::write_parallel("appSplit.vtk", &mesh, mesh.dim());
+    ts::writeVtk(mesh,"appSplit",0);
   }
   auto ptn = redev::ClassPtn(ranks,classIds);
   redev::Redev rdv(MPI_COMM_WORLD,ptn,isRdv);
@@ -226,13 +175,12 @@ int main(int argc, char** argv) {
 
   for(int iter=0; iter<3; iter++) {
     if(!rank) fprintf(stderr, "isRdv %d iter %d\n", isRdv, iter);
-    MPI_Barrier(MPI_COMM_WORLD);
     //////////////////////////////////////////////////////
     //the non-rendezvous app sends global vtx ids to rendezvous
     //////////////////////////////////////////////////////
     if(!isRdv) {
       //build dest, offsets, and permutation arrays
-      if(iter==0) prepareAppOutMessage(mesh, ptn, appOut, appOutPermute);
+      if(iter==0) ts::prepareAppOutMessage(mesh, ptn, appOut, appOutPermute);
       //fill message array
       auto gids = mesh.globals(0);
       auto gids_h = Omega_h::HostRead(gids);
@@ -252,7 +200,7 @@ int main(int argc, char** argv) {
       //attach the ids to the mesh
       if(iter==0) getRdvPermutation(mesh, rdvIn.msgs, rdvInPermute);
       ts::checkAndAttachIds(mesh, "inVtxGids", rdvIn.msgs, rdvInPermute);
-      writeVtk(mesh,"rdvInGids",iter);
+      ts::writeVtk(mesh,"rdvInGids",iter);
     } //end non-rdv -> rdv
     //////////////////////////////////////////////////////
     //the rendezvous app sends global vtx ids to non-rendezvous
